+++ StocksApp/Models/StockTrade.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace StocksApp.Models;
 
 /// <summary>
@@ -9,8 +7,6 @@ public class StockTrade
 {
     public string? StockSymbol { get; set; }
     public string? StockName { get; set; }
-    public double Price { get; set; } = 0;
+    public decimal Price { get; set; } = 0;
     public uint Quantity { get; set; } = 0;
 }
-
-
